+++ CHANGELOG.md
@@ -6,6 +6,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 
+Fixed:
+
+- Determining versions to migrate through (up/down).
 
 
 ## 0.2.0 - 2018-08-21